feat: consolidate record-replay functionality into unified package
Phase 1: Package restructuring and import updates

## Changes

### Package Structure
- Created unified `src/record-replay/` package with subdirectories:
  - `otlp-capture/` - Capture service (moved from src/otlp-capture)
  - `otlp-replay/` - Replay service (extracted)
  - `otlp-seed-generator/` - Placeholder for seed generation
  - `retention/` - Retention service (extracted)
  - `training/` - Training data reader (extracted)
  - `router/` - HTTP routers
  - `cli/` - CLI tools (placeholder)
  - `shared/` - Shared utilities (placeholder)
  - `test/` - Test organization

### File Moves (Git history preserved)
- Moved all otlp-capture files to record-replay/otlp-capture/
- Extracted replay-service.ts to otlp-replay/
- Extracted retention-service.ts to retention/
- Extracted training-data-reader.ts to training/
- Moved router.ts to router/capture-router.ts

### Import Updates
- Updated src/server.ts imports
- Updated src/annotations/router.ts imports
- Updated src/storage/router.ts imports
- Updated internal record-replay package imports
- Fixed all test file imports

### Testing
- All TypeScript compilation passes
- Test imports updated for new structure
- No functional changes to code

## Next Steps
- Phase 2: Implement HTTP client for replay
- Phase 3: Add seed generation with patterns
- Phase 4: Create CLI tools
- Phase 5: Update Docker configuration
- Phase 6: Delete old test-data generator

Issue #156
