+++ src/record-replay/otlp-capture/errors.ts
