+++ src/record-replay/otlp-capture/test/fixtures/otlp-generator.ts
