+++ src/record-replay/otlp-capture/test/integration/capture-replay.integration.test.ts
@@ -12,7 +12,7 @@ import {
   OtlpReplayServiceTag,
   OtlpReplayServiceLive
 } from '../../index.js'
-import { S3StorageTag } from '../../../storage/index.js'
+import { S3StorageTag } from '../../../../storage/index.js'
 import { generateTestOtlpData } from '../fixtures/otlp-generator.js'
 
 // This test uses testcontainers to spin up MinIO, so it should work in CI
@@ -74,7 +74,7 @@ describe('OTLP Capture and Replay Integration Tests', () => {
       S3StorageTag,
       Effect.gen(function* () {
         const { makeS3Storage } = yield* Effect.promise(() =>
-          import('../../../storage/s3.js')
+          import('../../../../storage/s3.js')
         )
         return yield* makeS3Storage(testS3Config)
       })
