+++ src/record-replay/otlp-capture/test/integration/retention-api.test.ts
@@ -8,8 +8,8 @@ import { Effect, Layer } from 'effect'
 import { GenericContainer, StartedTestContainer } from 'testcontainers'
 import { S3Client, CreateBucketCommand } from '@aws-sdk/client-s3'
 import { OtlpCaptureServiceTag, OtlpCaptureServiceLive } from '../../capture-service.js'
-import { RetentionServiceTag, RetentionServiceLive } from '../../retention-service.js'
-import { S3StorageTag, makeS3Storage } from '../../../storage/s3.js'
+import { RetentionServiceTag, RetentionServiceLive } from '../../../retention/retention-service.js'
+import { S3StorageTag, makeS3Storage } from '../../../../storage/s3.js'
 import { mockCaptureConfig, mockOtlpTraceData } from '../fixtures/test-data.js'
 import type { RetentionPolicy } from '../../index.js'
