+++ src/record-replay/otlp-capture/test/unit/errors.test.ts
